+++ src/app/(secured)/drawer.tsx
@@ -62,7 +62,7 @@ import { config } from '@/middleware';
 import { CloseOutlined } from '@mui/icons-material';
 //import { postServices } from '@/lib/redux';
 //import { useDispatch } from 'react-redux';
-import GoogleTranslatorPicker from '../../components/fullArticle/googleTranslator';
+import GoogleTranslatorPicker from '../../components/googleTranslate/googleTranslator';
 import { useDispatch, useSelector } from 'react-redux';
 import { ReduxState } from '@/lib/redux';
 //import SidebarProfileBar from '@/components/sidebarProfileInfo';
